v2.0
I have finished all of my features for sprint two.
Now the program can decide which direction to go in initially based off of the users startpoint, so that if the start point isnt placed on the right of a wall (eg. outside a mazes entrance where the entrance is open), then it will be able to go towards the maze and then operate as normal afterwards.